Create /lib/constants.js
