+++ lib/constants.js
@@ -1,0 +1,14 @@
+module.exports = {
+  API_ENDPOINTS: {
+    ALL_POST_IMAGES: `/api/posts/images`,
+    ALL_POSTS: `/api/posts/all`,
+    LOGIN: `/api/login`,
+    USER: `/api/user`,
+    CREATE_POSTS: `/api/posts/create/`,
+    DELETE_POSTS: `/api/posts/delete`,
+    UPDATE_POSTS: `/api/posts/update/`,
+    PAGE_VIEWS: `/api/page-views`,
+    IMAGES: `/api/images`,
+    IMAGE_UPLOAD: `/api/posts/uploadImageV2`
+  }
+}
